// apps/website/src/types/documentation.ts
export interface DocumentationClassMeta {
	line: number;
	file: string;
	path?: string;
	url?: string;
}

type DocumentationClassMethodMeta = DocumentationClassMeta;
type DocumentationClassMethodParameterMeta = DocumentationClassMeta;
type DocumentationClassMethodPropertyMeta = DocumentationClassMeta;
type DocumentationClassEventMeta = DocumentationClassMeta;
type DocumentationExternalMeta = DocumentationClassMeta;
type DocumentationTypeDefinitionMeta = DocumentationClassMeta;
type DocumentationTypeDefinitionParameterMeta = DocumentationClassMeta;

interface DocumentationParameter {
	name: string;
	description?: string;
	default: string;
	abstract?: boolean;
	deprecated?: boolean | string;
	access?: string;
	readonly?: boolean;
	optional?: boolean;
	variable?: boolean;
	nullable?: boolean;
	scope?: string;
	see?: string[];
	type: string[][][];
}
type DocumentationClassConstructorParameter = DocumentationParameter;
type DocumentationClassEventParameter = DocumentationParameter;

interface DocumentationClassMethodParameter extends DocumentationParameter {
	meta: DocumentationClassMethodParameterMeta;
}

interface DocumentationTypeDefinitionParameter extends DocumentationParameter {
	meta: DocumentationTypeDefinitionParameterMeta;
}

interface DocumentationProperty extends DocumentationParameter {}
type DocumentationTypeDefinitionProperty = DocumentationProperty;
type DocumentationClassPropertyProperty = DocumentationProperty;

export interface DocumentationClassProperty extends DocumentationProperty {
	props: DocumentationClassPropertyProperty[];
	meta: DocumentationClassMethodPropertyMeta;
}

interface DocumentationClassConstructor {
	name: string;
	params?: DocumentationClassConstructorParameter[];
}

export interface DocumentationClassMethod {
	name: string;
	description?: string;
	access?: string;
	async?: boolean;
	abstract?: boolean;
	deprecated?: boolean | string;
	inherited?: boolean;
	scope?: string;
	params?: DocumentationClassMethodParameter[];
	examples?: string[];
	returns?: unknown;
	returnsDescription?: string;
	throws?: string[];
	see?: string[];
	meta: DocumentationClassMethodMeta;
}